Add files via upload
This code implements the 1D-FC "blend to zero" matrix generation process as described by Bruno et al. in the following papers:
    
    https://arxiv.org/abs/2010.03901
    https://www.sciencedirect.com/science/article/abs/pii/S0021999115008086?via%3Dihub

It takes as inputs:
    d: The number of function values to sample; typically 1 <= d <= 12
    C: The number of continuation points to add.
    Z: The number of points in the "zero" region.
    E: The number of points following the zero region. Typically E = C
    n_over: The oversampling factor; typically n_over = 20
    modes_to_reduce: The number of Fourier modes to subtract from the least squares problem.
    num_digits: The number of digits to carry out the calculations with; default is 256.

And returns a single C x d matrix A, which can be used to compute the periodic continuation of a given sufficiently smooth, non-periodic data set. An example of how A is used is provided at the end of the code.

Note: While this computation can take a significant amount of time, the returned matrix A is not problem-dependent; once calculated and assuming the underlying calculation parameters (typically just d) work for a given problem, it can be saved and re-used as needed.

Note 2: While I authored this Python code, I am neither a co-author or inventor of FC. Please direct any theoretical questions about the underlying algorithm to the contacts specified in the above referenced papers.